requirements.txt
fastapi
pydantic
uvicorn
strawberry-graphql
supabase